refactor: refactor according to mentor's review
