+++ class-components/src/components/Element/Element.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, FC, useEffect, useState } from 'react';
+import { ChangeEvent, FC, useMemo } from 'react';
 import { Link } from 'react-router-dom';
 import { getElementId } from '../../helpers/getElementId';
 import useAppSelector from '../../hooks/useAppSelector';
@@ -19,17 +19,13 @@ const Element: FC<Props> = ({ person }) => {
   const { url, name } = person;
   const dispatch = useAppDispatch();
   const { selectedPeople } = useAppSelector((state) => state.people);
-  const index = selectedPeople.findIndex((element) => element.url === url);
-  const [isSelected, setIsSelected] = useState(index !== -1);
+  const isSelected = useMemo(
+    () => Boolean(selectedPeople.find((element) => element.url === url)),
+    [selectedPeople, url],
+  );
   const id = getElementId(url);
 
-  useEffect(() => {
-    const index = selectedPeople.findIndex((element) => element.url === url);
-    setIsSelected(index !== -1);
-  }, [selectedPeople, url]);
-
   function changeHandler({ target }: ChangeEvent<HTMLInputElement>) {
-    setIsSelected(target.checked);
     if (target.checked) {
       dispatch(addSelectedPerson(person));
     } else {
