+++ class-components/src/components/ThemeControls/ThemeControls.tsx
@@ -4,8 +4,14 @@ import { ThemeContext } from '../../helpers/context';
 
 import './themeControls.css';
 
+enum Icons {
+  sun = 'sun',
+  moon = 'moon',
+}
+
 const ThemeControls: FC = () => {
   const { isDark, setIsDark } = useContext(ThemeContext) as ThemeContextType;
+  const icon = isDark ? Icons.moon : Icons.sun;
 
   function changeTheme() {
     setIsDark(!isDark);
@@ -14,11 +20,7 @@ const ThemeControls: FC = () => {
   return (
     <div className="theme-controls">
       <button onClick={changeTheme}>
-        {isDark ? (
-          <img src="/moon.svg" alt="moon icon" />
-        ) : (
-          <img src="/sun.svg" alt="sun icon" />
-        )}
+        <img src={`/${icon}.svg`} alt={`${icon} icon`} />
       </button>
     </div>
   );
